refactor: .js 파일 ts로 변경
